requirements.txt
django>=4.0, <5.0
gunicorn
django-heroku
Pillow
